requirements.txt
Django
graphene-django >= 2.0
psycopg2-binary
whitenoise
graphene-file-upload
Pillow